+++ src/test/kotlin/apply/application/MailServiceTest.kt
@@ -2,8 +2,10 @@ package apply.application
 
 import apply.createMailData
 import apply.createMailHistory
+import apply.createUser
 import apply.domain.mail.MailHistoryRepository
 import apply.domain.user.UserRepository
+import apply.domain.user.findByEmail
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import org.assertj.core.api.Assertions.assertThat
@@ -11,6 +13,7 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 import support.test.UnitTest
+import support.views.NO_NAME
 import java.time.LocalDateTime
 
 @UnitTest
@@ -44,5 +47,21 @@ class MailServiceTest {
         val emailHistory2 = createMailHistory(subject = "제목2", sentTime = now.plusSeconds(1))
         every { mailHistoryRepository.findAll() } returns listOf(emailHistory1, emailHistory2)
         assertThat(mailService.findAll()).containsExactly(mailData1, mailData2)
+    }
+
+    @Test
+    fun `회원 중에서 해당하는 이메일이 있으면 이름과 이메일을 가져온다`() {
+        val user = createUser()
+        every { userRepository.findByEmail(user.email) } returns user
+        val actual = mailService.findMailTargetByEmail(user.email)
+        assertThat(actual.name).isEqualTo(user.name)
+    }
+
+    @Test
+    fun `회원 중에서 해당하는 이메일이 없으면 (이름없음)과 이메일을 반환한다`() {
+        val email = "email@email.com"
+        every { userRepository.findByEmail(email) } returns null
+        val actual = mailService.findMailTargetByEmail(email)
+        assertThat(actual.name).isEqualTo(NO_NAME)
     }
 }
